UP: Add img and card for pizza

+++ la-mia-pizzeria-static/Models/Pizze.cs
@@ -5,11 +5,11 @@
         public string Name { get; set; }
         public string Description { get; set; }
         public string Img { get; set; }
-        public float Price { get; set; }
+        public double Price { get; set; }
 
         public Pizze() { }
 
-        public Pizze(string name, string description, string img, float price)
+        public Pizze(string name, string description, string img, double price)
         {
             this.Name = name;
             this.Description = description;
